Added Safeguards to DigitalOcean Commands

+++ control.sh
@@ -60,14 +60,6 @@ attach)
     docker attach ngradar_website_service
     ;;
 
-# Now that our live db lives in a DO droplet, I don't think we can run this command unless we want to ssh into the 
-# dsoc droplet. but I don't think we use this command enough for it to matter since all the data in the live
-# db is dummy data anyways. 
-# load-staging-data)
-#     docker compose run --rm staging_loader
-#     ;;
-
-
 kafka-up)
     echo "Starting Kafka infrastructure and storage..."
     docker compose $KAFKA_PROFILES up -d $KAFKA_SERVICES
@@ -84,7 +76,6 @@ system-up)
     echo "Starting simulator services..."
     "$0" sims-up
     ;;
-
 
 kafka-down)
     echo "Stopping kafka infrastructure and storage..."
@@ -132,7 +123,6 @@ testcov)
     ;;
 
 hard-reset)
-
     read -p "This will DELETE your local database and containers. Continue? (y/N): " ANSWER
 
     if [[ "$ANSWER" != "y" && "$ANSWER" != "Y" ]]; then
@@ -155,15 +145,34 @@ agent-up)
     ;;
 
 gbt-up)
+    read -p "This will START the gbt container when run in the GBT DIGITALOCEAN DROPLET. Continue? (y/N): " ANSWER
+
+    if [[ "$ANSWER" != "y" && "$ANSWER" != "Y" ]]; then
+        exit 0
+    fi
+
     docker compose up -d $GBT_SERVICES
     ;;
 
 gbt-down)
+    read -p "This will STOP the gbt container when run in the GBT DIGITALOCEAN DROPLET. Continue? (y/N): " ANSWER
+
+    if [[ "$ANSWER" != "y" && "$ANSWER" != "Y" ]]; then
+        exit 0
+    fi
+
     docker compose stop $GBT_SERVICES
     docker compose rm -f $GBT_SERVICES
     ;;
 
 vlba-up)
+    # TODO change this line when scaling up
+    read -p "This will START the vlba container when run in the VLBA DIGITALOCEAN DROPLET. Continue? (y/N): " ANSWER
+
+    if [[ "$ANSWER" != "y" && "$ANSWER" != "Y" ]]; then
+        exit 0
+    fi
+
     # TODO start all of these when scaling up!
     # Get the services that are passed as the argument to vlba-up
     SERVICES="${!2}"
@@ -171,6 +180,13 @@ vlba-up)
     ;;
 
 vlba-down)
+    # TODO change this line when scaling up
+    read -p "This will STOP the vlba container when run in the VLBA DIGITALOCEAN DROPLET. Continue? (y/N): " ANSWER
+
+    if [[ "$ANSWER" != "y" && "$ANSWER" != "Y" ]]; then
+        exit 0
+    fi
+
     # TODO remove all of these when scaling up!
     # Get the services that are passed as the argument to vlba-down
     SERVICES="${!2}"
@@ -179,15 +195,33 @@ vlba-down)
     ;;
 
 dsoc-up)
+    read -p "This will START the dsoc container when run in the DSOC DIGITALOCEAN DROPLET. Continue? (y/N): " ANSWER
+
+    if [[ "$ANSWER" != "y" && "$ANSWER" != "Y" ]]; then
+        exit 0
+    fi
+
     docker compose up -d $DSOC_SERVICES
     ;;
 
 dsoc-down)
+    read -p "This will STOP the dsoc container when run in the DSOC DIGITALOCEAN DROPLET. Continue? (y/N): " ANSWER
+
+    if [[ "$ANSWER" != "y" && "$ANSWER" != "Y" ]]; then
+        exit 0
+    fi
+
     docker compose stop $DSOC_SERVICES
     docker compose rm -f $DSOC_SERVICES
     ;;
 
 droplets-up)
+    read -p "Are you sure you want to ssh into the DIGITALOCEAN DROPLETS? This will START all of the DIGITALOCEAN DROPLETS from your LOCAL TERMINAL. Continue? (y/N): " ANSWER
+
+    if [[ "$ANSWER" != "y" && "$ANSWER" != "Y" ]]; then
+        exit 0
+    fi
+
     echo "Starting DSOC Droplet"
     ssh "$DSOC_DROPLET" \
         "cd $REMOTE_DIR && ./control.sh dsoc-up"
@@ -211,6 +245,12 @@ droplets-up)
     ;;
 
 droplets-down)
+    read -p "Are you sure you want to ssh into the DIGITALOCEAN DROPLETS? This will STOP all of the DIGITALOCEAN DROPLETS from your LOCAL TERMINAL. Continue? (y/N): " ANSWER
+
+    if [[ "$ANSWER" != "y" && "$ANSWER" != "Y" ]]; then
+        exit 0
+    fi
+
     echo "Stopping GBT Droplet"
     ssh "$GBT_DROPLET" \
         "cd $REMOTE_DIR && ./control.sh gbt-down"
@@ -261,8 +301,8 @@ droplets-down)
     echo "./control.sh dsoc-down"
     echo "./control.sh gbt-up"
     echo "./control.sh gbt-down"
-    echo "./control.sh vlba-up"
-    echo "./control.sh vlba-down"
+    echo "./control.sh vlba-up VLBA_#_SERVICES"
+    echo "./control.sh vlba-down VLBA_#_SERVICES"
     echo "./control.sh portainer-up"
     echo "./control.sh agent-up"
     echo
